+++ src/changelog.md
@@ -1,3 +1,9 @@
+#0.1.5
+  * Added: context menu dropdown in modal
+  * Added: action to clear all data from storage
+  * Added: action to clear only current data from storage
+  * Added: action to log all storage data to console
+  * Updated screens
 #0.1.1
   * Start rewrite app
   * Now added Chrome synchronization